Move platform implementations to subfolder
